requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
Pillow==6.0.0
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.1
sqlparse==0.3.0
